common run config to separet file
