+++ package.json
@@ -13,14 +13,16 @@
     "test:node": "cd api && yarn test",
     "test:smoke": "sh whitelist.sh && ts-node ./test/integration/tests/test.ts",
     "test:smoke:functional": "sh whitelist.sh && webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/smoke.conf.js --local",
-    "test:functional": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/conf.js --local",
+    "test:functional": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/functional.conf.js",
+    "test:functional:local": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/functional.conf.js --local",
+    "test:fullfunctional": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/fullfunctional.conf.js",
+    "test:fullfunctional:local": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/fullfunctional.conf.js --local",
     "test:coverage": "echo 'not implemented'",
     "test:node:watch": "cd api && yarn test:watch",
     "test:api": "ts-node ./test/integration/tests/test.ts",
     "test:coverage:ng": "ng test rpa-pui-manager --code-coverage --watch",
     "test:coverage:node": "cd api && yarn coverage",
     "test:smoke:local": "webdriver-manager update --versions.chrome 2.40 && protractor ./test/e2e/config/smoke.conf.js --local",
-    "test:functional:local": "protractor ./test/e2e/config/functional.conf.js --local",
     "ng": "ng",
     "start:ng": "ng serve --port 3000 --host 0.0.0.0 --proxy-config proxy.config.json",
     "build:ng": "ng build",
