+++ test/e2e/config/smoke.conf.js
@@ -92,7 +92,7 @@ const config = {
     strict: true,
     // format: ['node_modules/cucumber-pretty'],
     format: ['node_modules/cucumber-pretty', 'json:reports_json/results.json'],
-    tags: ['@ignore'],
+    tags: ['@smoke'],
     require: [
       '../support/timeout.js',
       '../support/world.js',
