+++ test/e2e/features/app/createOrganisation.feature
@@ -1,5 +1,4 @@
 
-@all
 Feature: register organization
 
   Background:
