+++ test/e2e/features/app/loginLogout.feature
@@ -1,4 +1,4 @@
-
+@all
 Feature: Login
 
   Background:
@@ -18,27 +18,23 @@ Feature: Login
 #    Then I logout successfully and back to Login page
 
 
-  @all
+
   Scenario: un-authenticated user login
     Then I am on Idam login page
     When I enter an Invalid email-address and password to login
     Then I should be redirected to the Idam login page
     Then I should see failure error summary
 
+
+@smoke 
+  Scenario: login and log out from manage organisation as ManageOrg user
+    Given I am logged into manage organisation with ManageOrg user details
+    Then I should be redirected to manage organisation dashboard page
+    When I select the sign out link
+    Then I should be redirected to the Idam login page
+
  
-  Scenario: login and log out from manage organisation as ManageOrg user
-    Given I am logged into manage organisation with ManageOrg user details
-    Then I should be redirected to manage organisation dashboard page
-    When I select the sign out link
-    Then I should be redirected to the Idam login page
-
-  Scenario: login and log out from manage organisation as ManageOrg user
-    Given I am logged into manage organisation with ManageOrg user details
-    Then I should be redirected to manage organisation dashboard page
-    When I select the sign out link
-    Then I should be redirected to the Idam login page
-
-  @all
+  @smoke
   Scenario: Verify the direct link navigate to login page
     Given I navigate to manage organisation Url direct link
     Then I should be redirected back to Login page after direct link
