+++ test/e2e/features/step_definitions/loginLogout.steps.js
@@ -3,7 +3,7 @@
 const loginPage = require('../pageObjects/loginLogoutObjects');
 const { defineSupportCode } = require('cucumber');
 const { AMAZING_DELAY, SHORT_DELAY, MID_DELAY, LONG_DELAY } = require('../../support/constants');
-const config = require('../../config/conf.js');
+const config = require('../../config/common.conf.js');
 const EC = protractor.ExpectedConditions;
 
 const mailinatorService = require('../pageObjects/mailinatorService');
