+++ test/e2e/features/step_definitions/viewOrganisation.steps.js
@@ -2,7 +2,6 @@ let ViewOrganisationPage = require('../pageObjects/viewOrganisationPage.js');
 const headerPage = require('../pageObjects/headerPage');
 const loginPage = require('../pageObjects/loginLogoutObjects');
 const { AMAZING_DELAY, SHORT_DELAY, MID_DELAY, LONG_DELAY } = require('../../support/constants');
-const config = require('../../config/conf.js');
 const EC = protractor.ExpectedConditions;
 
 var {defineSupportCode} = require('cucumber');
