+++ test/e2e/support/hooks.js
@@ -3,7 +3,7 @@ const Cucumber = require('cucumber');
 const { defineSupportCode } = require('cucumber');
 const fs = require('fs');
 const mkdirp = require('mkdirp');
-const conf = require('../config/conf').config;
+const conf = require('../config/common.conf').config;
 // const conf = require('../config/saucelabs.conf').config;
 const reporter = require('cucumber-html-reporter');
 const report = require('cucumber-html-report');
